+++ src/client.ts
@@ -3,7 +3,7 @@
 //
 
 export * from "./components/forms/Form.client.js";
-export * from "./components/forms/PasswordInput.client.js";
+export * from "./components/forms/Input.client.js";
 
 export * from "./components/HumanDateTime.client.js";
 export * from "./components/HumanRelativeTime.client.js";