+++ src/index.ts
@@ -2,21 +2,11 @@
 // Exports
 //
 
-export * from "./components/forms/CheckboxInput.js";
-export * from "./components/forms/ColorInput.js";
-export * from "./components/forms/DateInput.js";
-export * from "./components/forms/EmailInput.js";
-export * from "./components/forms/Fieldset.js";
-export * from "./components/forms/FileInput.js";
+export * from "./components/forms/Checkbox.js";
 export * from "./components/forms/Form.js";
 export * from "./components/forms/HiddenInput.js";
+export * from "./components/forms/Input.js";
 export * from "./components/forms/Label.js";
-export * from "./components/forms/NumberInput.js";
-export * from "./components/forms/PasswordInput.js";
-export * from "./components/forms/RadioInput.js";
-export * from "./components/forms/Select.js";
-export * from "./components/forms/TextArea.js";
-export * from "./components/forms/TextInput.js";
 
 export * from "./components/head/AppleTouchIcon.js";
 export * from "./components/head/ICOFavicon.js";